+++ src/test/java/be/ugent/rml/Mapper_MySQL_Test.java
@@ -1,5 +1,6 @@
 package be.ugent.rml;
 
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -91,6 +92,7 @@ public class Mapper_MySQL_Test extends MySQLTestCore {
     }
 
     @Test
+    @Ignore
     public void doMapping() throws Exception {
         //setup expected exception
         if (expectedException != null) {